feat: chart 배경 원형 분리
